+++ examples/with-sveltekit/eslint.config.mjs
@@ -23,9 +23,9 @@ export default [
   },
   js.configs.recommended,
   ...ts.configs.recommended,
-  ...svelte.configs['flat/recommended'],
+  ...svelte.configs.recommended,
   prettier,
-  ...svelte.configs['flat/prettier'],
+  ...svelte.configs.prettier,
   {
     languageOptions: {
       globals: {